added bottom sheet for details (no details yet) (#7)
* added bottom sheet for details (no details yet)

also removed arFragment in favor of arSceneView because bottom sheet didn't
play well with fragment

* review fixes

+++ app/src/main/kotlin/co/netguru/android/carrecognition/data/recognizer/TFlowRecognizer.kt
@@ -63,7 +63,7 @@ class TFlowRecognizer @Inject constructor(private val tflow: Interpreter,
                     .mapIndexed { index, confidence ->
                         Recognition(
                             Car.of(labels[index]),
-                            confidence.toDouble() / Byte.MAX_VALUE
+                            confidence.toFloat() / Byte.MAX_VALUE
                         )
                     }
                     .sortedByDescending { it.confidence }
